requirements.txt
fastapi==0.85.0
mysql_connector_repackaged==0.3.1
pydantic==1.10.2
pytz==2021.1